social media links using web4 (#111)
This PR adds the `web4_get` method to the contract, so that when merged,
there will be a website at https://devhub.near.page

You can preview it at https://devhublink.testnet.page/

<img width="612" alt="image"
src="https://github.com/NEAR-DevHub/neardevhub-contract/assets/9760441/c0457ee5-df36-4549-97c2-a6a1a5fb77b3">

Note that the preview deployment does not have access to content, so it
will not be able to display proper content for preview links, which is
provided by the contract. A test community "webassemblymusic" is created
for the preview deployment, to demonstrate the preview links. You can
test it at https://devhublink.testnet.page/community/webassemblymusic.
You may also call the preview contract to create proposals and other
communities to verify that social media headers are also populated with
the content.

For now the first part of the part is always mapped to the `page`
property, so that e.g. `/proposals` are translated to
`/devhub.near/widget/app?page=proposals`. Also the second part
translates to `handle` for community ( `/community/webassemblymusic` ->
`/devhub.near/widget/app?page=community&handle=webassemblymusic` ), or
to `id` for proposal `/proposal/5` ->
`/devhub.near/widget/app?page=proposal&id=5`

In addition to metadata for social media links, and SEO friendly URLs,
the actual content is viewed as well. This is done using
[near-bos-webcomponent](https://github.com/NEARBuilders/near-bos-webcomponent/pull/10)
which is also used for playwright tests in the DevHub frontend. A PR
have also been created there to support the SEO friendly URL paths
created here.

The `html-escape` crate is added to ensure that descriptions with html
tags are escaped for the generated HTML ( to avoid malicious scripts etc
).

Examples:

- https://devhublink.testnet.page/community/webassemblymusic ( This
community is also created in the testnet contract )
- https://devhublink.testnet.page/community/chain-abstraction ( not
created in the testnet contract, so metadata is missing, but the content
shows )
- https://devhublink.testnet.page/proposal/0 ( Also has a proposal
created in the testnet contract )
- https://devhublink.testnet.page/proposal/3 ( not created in the
testnet contract, so metadata is missing, but the content shows )

Part of https://github.com/NEAR-DevHub/neardevhub-bos/issues/724


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced a new `web4` module to handle web requests, dynamically
generate HTML content, and encode responses.
- Added a public function for fetching web content based on request
paths, enhancing user interaction with community and proposal pages.

- **Documentation**
	- Enhanced documentation for new web request handling functionality.

- **Dependencies**
	- Added `html-escape` library to support HTML content generation.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: Vlad Frolov <frolvlad@gmail.com>