feat: additional kyc check for review stage (#138)
Closes https://github.com/NEAR-DevHub/neardevhub-bos/issues/848 - adds
additional kyc check for review stage.

For that, VersionedTimelineStatus along with TimelineStatusV2 is
introduced

Also it removes posts functionality for the following reason:
```bash
Action #0: ExecutionError("Exceeded the prepaid gas.")
```
It arises on calling either `unsafe_self_upgrade` or `unsafe_migrate` on
`test_deploy_contract_self_upgrade` test. Because code size is too big. So
code for posts was removed.