Web4 scripts from socialdb (#148)
Preview URL: https://devhubweb4.testnet.page/

Notice that the served index.html is referencing `web4browserclient.js`
with a `blockHeight` parameter. The block height of the latest deployed
`web4browserclient.js` is obtained to a preload call to socialdb. This
ensures that the browser will always load the latest version of the
script, and allow local caching as long as there are no changes.

<img width="1249" alt="image"
src="https://github.com/user-attachments/assets/468e78a2-7393-4fd0-9db3-e4c55ac1a6c5">

The javascript file, which is stored in socialdb, can then be loaded
from this URL:
https://devhubweb4.testnet.page/resources/devhub.near/web4browserclient.js?blockHeight=127582922

This will query socialdb for the latest block height of the javascript
referenced in the web4 deployment, and reference the script in the
served index.html. The script will also served through web4. Having the
blockheight in the query string when referencing the script, ensure that
it's always the latest script being loaded, but also enables caching in
the browser as long as there are no changes. This is a good alternative
to referencing a CDN where you often don't have the same opportunity to
check which version is the latest, and also we can reduce the number of
dependencies this way. Both when it comes to referencing the script, as
we now only can use the web4 gateway, and also in deployment, where
everything can be deployed to socialdb.

Resolves #146